Revert "no message"
This reverts commit b6106e8a959e09007ffdb0de8851b6f4b35eec3b.

+++ CIGA/Assets/Scripts/DarkEffect.cs
@@ -13,7 +13,7 @@ public class DarkEffect : MonoBehaviour
         public Transform target;
 
         [SerializeField]
-        public float radius;
+        public int radius;
 
         public Vector3 GetScreenPosition(Camera cam)
         {
@@ -38,7 +38,7 @@ public class DarkEffect : MonoBehaviour
     Vector4 _tmpVt;
     Vector3 _tmpPos;
     int _tmpScreenHeight;
-    public void setRoleDis(float val)
+    public void setRoleDis(int val)
     {
         _items[0].radius = val;
     }
